добавила описание моделей
