+++ MoviesMVVM/MoviesMVVM/MainModule/Model/Genres.swift
@@ -1,0 +1,10 @@
+// Genres.swift
+// Copyright © RoadMap. All rights reserved.
+
+import Foundation
+
+/// Модель жанров
+struct Genres: Decodable {
+    /// Жанр
+    let name: String
+}
