+++ MoviesMVVM/MoviesMVVM/MainModule/Model/Result.swift
@@ -1,0 +1,17 @@
+// Result.swift
+// Copyright © RoadMap. All rights reserved.
+
+import Foundation
+
+/// Массив с фильмами
+struct Result: Decodable {
+    /// Фильмы
+    let filmsInfo: [Movie]
+    /// Количество страниц
+    let pageCount: Int
+
+    enum CodingKeys: String, CodingKey {
+        case filmsInfo = "results"
+        case pageCount = "total_pages"
+    }
+}
