+++ MoviesMVVM/MoviesMVVM/MainModule/Model/VideoId.swift
@@ -5,10 +5,12 @@ import Foundation
 
 /// Модель массива информации о видео
 struct ResultVideos: Decodable {
+    /// Ссылки на видео тизеров
     let results: [VideoId]
 }
 
 /// Ключ от видео
 struct VideoId: Decodable {
+    /// Часть ссылки на страницу с фильмом
     let key: String
 }
